requirements.txt
dynaconf>=3.1.11
tgtg>=0.14.0
schedule>=1.1.0
requests>=2.28.2
python-dateutil>=2.8.1
pytest>=7.2.1
pylint>=2.16.3
